# db/internal: use bytes.Buffer