```
benchmark                         old ns/op     new ns/op     delta
BenchmarkMakeARRAYOfBIGINTs-2     26671         4559          -82.91%

benchmark                         old allocs     new allocs     delta
BenchmarkMakeARRAYOfBIGINTs-2     300            9              -97.00%

benchmark                         old bytes     new bytes     delta
BenchmarkMakeARRAYOfBIGINTs-2     22770         2040          -91.04%
```