@chappjc 2019-03-07
Nice.  Just a couple comments.  strings.Builder would be even better about allocs.  Also, I'm not certain we're using all these functions.