@chappjc 2019-03-08
Wanna check in those benchmarks?  Would be helpful.